Update README.md
Catman lives in a small town. He is the cop of the town. Sometimes he is attracted by the fish in the fish tank while working. He always wants to catch fish. He likes to eat fish. Sometimes when he was dealing with other things in the town, he would meet super sisters. These sisters could always help him when he needed help the most. catman is a brave and kind chief and everyone loves him